+++ packages/class-core-test/src/value-managers/property_0_1.tests.ts
@@ -23,7 +23,7 @@ import {
     PropertyDeletedDelta,
     serializeNodeBases
 } from "@lionweb/class-core"
-import { AccumulatingSimplisticHandler, DefaultPrimitiveTypeDeserializer } from "@lionweb/core"
+import { AccumulatingSimplisticHandler, BuiltinPropertyValueDeserializer } from "@lionweb/core"
 import { LionWebJsonMetaPointer } from "@lionweb/json"
 
 import { deepEqual, equal, isTrue } from "../assertions.js"
@@ -62,7 +62,7 @@ describe("[0..1] string property", () => {
         const [handleDelta, deltas] = collectingDeltaHandler();
         const deserialize = nodeBaseDeserializer([testLanguageBase], handleDelta);
         const problemHandler = new AccumulatingSimplisticHandler();
-        const deserializedNodes = deserialize(serializationChunk, [], new DefaultPrimitiveTypeDeserializer(), problemHandler);
+        const deserializedNodes = deserialize(serializationChunk, [], new BuiltinPropertyValueDeserializer(), problemHandler);
         equal(problemHandler.allProblems.length, 0);
         equal(deserializedNodes.length, 1);
         const root = deserializedNodes[0];
@@ -110,7 +110,7 @@ describe("[0..1] string property", () => {
         const [handleDelta, deltas] = collectingDeltaHandler();
         const deserialize = nodeBaseDeserializer([testLanguageBase], handleDelta);
         const problemHandler = new AccumulatingSimplisticHandler();
-        const deserializedNodes = deserialize(serializationChunk, [], new DefaultPrimitiveTypeDeserializer(), problemHandler);
+        const deserializedNodes = deserialize(serializationChunk, [], new BuiltinPropertyValueDeserializer(), problemHandler);
         equal(problemHandler.allProblems.length, 0);
         equal(deserializedNodes.length, 1);
         const root = deserializedNodes[0];